separated Game as a component, and moved utils to its appropriate components' files. Issue #6
