+++ src/utils/scoreBoardUtils.js
@@ -1,41 +1,3 @@
-import { setNewScoreThunk } from '../redux/actions.js';
-
-const sortScores = (a, b) => {
-    const scoreA = parseInt(a['Score']);
-    const scoreB = parseInt(b['Score']);
-
-    if (scoreA > scoreB) return -1;
-    if (scoreA === scoreB) return 0;
-    if (scoreA < scoreB) return 1;
-};
-
-
-export const getTopSortedScores = (scoreBoardResults, howMany) => {
-    const scoreBoardArray = Object.entries(scoreBoardResults);
-
-    const scores = [];
-    scoreBoardArray.forEach(
-        row =>
-            scores.push({
-                'User': row[0],
-                'Score': parseInt(row[1]['Score']),
-                'RowKey' : row[1]['TimeStamp']
-            })
-    );
-
-    scores.sort(sortScores);
-    return scores.slice(0, howMany);
-}
-
-export const getOnlyScores = (scoresObj) => {
-    const onlyScores = [];
-    scoresObj.forEach(
-        row => onlyScores.push(row['Score'])
-    );
-
-    return onlyScores;
-}
-
 export const getRowElements = (topScores, Score) => {
     const rows = [];
 
@@ -49,7 +11,6 @@ export const getRowElements = (topScores, Score) => {
 
     return rows;
 }
-
 
 const getCountOfMinNums = (topScoresArr) => {
     const minNum = Math.min(...topScoresArr);
@@ -140,29 +101,4 @@ export const checkForNewRecord = (score, topScoresArray, boardLength) => {
     }
 
     return [newRecordSet, whichIndex];
-}
-
-
-export const handleSubmit = (event, userName, whichIndex, itemAtIndex, score, dispatch) => {
-    event.preventDefault();
-
-    if (userName.trim().length === 0)
-        alert('You can\'t submit a blank form!');
-
-    // run through regex
-    const isOnlyStandardChars = /^[a-zA-Z1-9-_*$]+$/.test(userName);
-
-    if (!isOnlyStandardChars) {
-        alert('You can only use a-z, A-Z, 1-9, -, _ , * ,$ characters.');
-    } else {
-        const confirmUserName = prompt(`Submitting you score under userName: ${userName}? (yes/no)`);
-
-        if (confirmUserName === null) {
-            alert('You are a picky one!\nChange your name or restart the game!')
-
-        } else if (confirmUserName.toLowerCase() === 'yes') {
-            alert('sweet!');
-            dispatch(setNewScoreThunk(userName, score, 'sergey', 10));
-        }
-    }
 }
